tensorgen/pyproject.toml
[project]
name = "tensorgen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.2",
    "pillow>=11.3.0",
    "safetensors>=0.5.3",
    "torch>=2.7.1",
]
